# Add native `list_has_any` implementation