--- a/src/core_functions/scalar/list/list_has_any.cpp
+++ b/src/core_functions/scalar/list/list_has_any.cpp
@@ -1,8 +1,8 @@
 #include "duckdb/core_functions/lambda_functions.hpp"
 #include "duckdb/core_functions/scalar/list_functions.hpp"
+#include "duckdb/core_functions/create_sort_key.hpp"
 #include "duckdb/planner/expression/bound_cast_expression.hpp"
-
-#include <duckdb/common/string_map_set.hpp>
+#include "duckdb/common/string_map_set.hpp"
 
 namespace duckdb {