This PR implements the `list_has_any` function natively instead of having it be provided by a default macro. 

I haven't run any benchmarks, but this implementation "sort-key" serializes both lists and hashes the smaller list into a `string_set_t` before probing with the larger list, which should at least bring down the theoretical complexity from `n * m` to `n + m`, even if the sort-key serialization has some overhead.

The main motivation for implementing this as a function instead of a macro is to free up the `&&` alias for use in extensions (spatial) as macros can't be overloaded (or shadowed by a function as macros seem to bind stronger).