@Maxxen 2024-07-26
Here's a benchmark!
```sql
load
CREATE TABLE t1 as SELECT range(s, s + 3000) as l1, range(s + 1499, s + 3000) as l2 FROM range(0, 3000) r(s);

run
SELECT bool_and(list_has_any(l1, l2)) FROM t1;

result I
true
```

Main:
```
benchmark/micro/list/list_has_any.benchmark     1       2.907506
benchmark/micro/list/list_has_any.benchmark     2       2.912112
benchmark/micro/list/list_has_any.benchmark     3       2.677297
benchmark/micro/list/list_has_any.benchmark     4       2.912805
benchmark/micro/list/list_has_any.benchmark     5       2.672608
````

With this PR:
```
benchmark/micro/list/list_has_any.benchmark     1       0.269804
benchmark/micro/list/list_has_any.benchmark     2       0.269562
benchmark/micro/list/list_has_any.benchmark     3       0.270481
benchmark/micro/list/list_has_any.benchmark     4       0.273571
benchmark/micro/list/list_has_any.benchmark     5       0.261554
```